Refactor Gemini logic into service and split screen
